% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdmurl.R
\name{as.tdmurl}
\alias{as.tdmurl}
\title{Coerce a url to a tdmurl with a specific type}
\usage{
as.tdmurl(...)
}
\description{
Coerce a url to a tdmurl with a specific type
}
\note{
see \code{crminer::as_tdmurl}
}
\keyword{internal}
