% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cr_agency}
\alias{cr_agency}
\title{Check the DOI minting agency on one or more dois}
\usage{
cr_agency(dois = NULL, .progress = "none", ...)
}
\arguments{
\item{dois}{(character) One or more article or organization dois.}

\item{.progress}{Show a \code{plyr}-style progress bar? Options are "none", "text", "tk", "win,
and "time".  See \link[plyr]{create_progress_bar} for details of each.}

\item{...}{Named parameters passed on to httr::GET}
}
\description{
Check the DOI minting agency on one or more dois
}
\details{
See \url{bit.ly/1nIjfN5} for more info on the Fundref API service.
}
\examples{
\dontrun{
cr_agency(dois = '10.13039/100000001')
res <- cr_agency(dois = c('10.13039/100000001','10.13039/100000015'))
res[['10.13039/100000015']]
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

