% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cr_citation_count}
\alias{cr_citation_count}
\title{Lookup article info via CrossRef with DOI and get a citation count.}
\usage{
cr_citation_count(doi, url = "http://www.crossref.org/openurl/",
  key = "cboettig@ropensci.org", ..., curl = getCurlHandle())
}
\arguments{
\item{doi}{digital object identifier for an article in PLoS Journals}

\item{url}{the PLoS API url for the function (should be left to default)}

\item{key}{your PLoS API key, either enter, or loads from .Rprofile}

\item{...}{optional additional curl options (debugging tools mostly)}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\value{
citation count
}
\description{
Lookup article info via CrossRef with DOI and get a citation count.
}
\details{
See \url{http://labs.crossref.org/openurl/} for more info on this
		Crossref API service.
}
\examples{
\dontrun{
cr_citation_count(doi="10.1371/journal.pone.0042793")
}
}
\author{
Carl Boettiger \email{cboettig@gmail.com}
}
\seealso{
\code{\link{cr_search}}, \code{\link{cr_r}}, \code{\link{cr_search_free}}
}

