% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cr_cn}
\alias{cr_cn}
\title{Get citations in various formats from CrossRef.}
\usage{
cr_cn(dois, format = "text", style = "apa", locale = "en-US",
  .progress = "none", ...)
}
\arguments{
\item{dois}{Search by a single DOI or many DOIs.}

\item{format}{Name of the format. One of "rdf-xml", "turtle", "citeproc-json", "text",
"ris", "bibtex", "crossref-xml", "datacite-xml", or "bibentry"}

\item{style}{a CSL style (for text format only). See \code{\link{get_styles}}
for options. Default: apa. If there's a style that CrossRef doesn't support you'll get a
\code{(500) Internal Server Error}}

\item{locale}{Language locale. See \code{?Sys.getlocale}}

\item{.progress}{Show a \code{plyr}-style progress bar? Options are "none", "text", "tk", "win,
and "time".  See \code{\link[plyr]{create_progress_bar}} for details of each.}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Get citations in various formats from CrossRef.
}
\details{
See \url{http://www.crosscite.org/cn/} for more info on the
  	Crossref Content Negotiation API service.
}
\examples{
\donttest{
cr_cn(dois="10.1126/science.169.3946.635")
}
}

