% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cr_full.R
\name{cr_ft_links}
\alias{cr_ft_links}
\title{Get full text links from a DOI}
\usage{
cr_ft_links(doi, type = "xml", ...)
}
\arguments{
\item{doi}{A DOI}

\item{type}{One of xml, plain, pdf, or all}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Get full text links from a DOI
}
\details{
Note that \code{\link{cr_ft_links}} is not vectorized.
}
\examples{
\dontrun{
# pdf link
cr_ft_links(doi = "10.5555/515151", "pdf")

# xml and plain text links
out <- cr_works(filter=c(has_full_text = TRUE))
dois <- out$data$DOI
cr_ft_links(dois[2], "xml")
cr_ft_links(dois[1], "plain")
cr_ft_links(dois[1], "all")

# No links
cr_ft_links(cr_r(1), "xml")
}
}

