% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_ft_text.R
\name{cr_ft_plain}
\alias{cr_ft_plain}
\title{Get full text plain from a DOI}
\usage{
cr_ft_plain(...)
}
\description{
Get full text plain from a DOI
}
\note{
see \code{crminer::crm_plain}
}
\keyword{internal}
