% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cr_full}
\alias{cr_full}
\title{Get full text links from a DOI}
\usage{
cr_full(doi, type = "xml", ...)
}
\arguments{
\item{doi}{A DOI}

\item{type}{One of xml, plain, pdf, or all}

\item{...}{Further curl args}
}
\description{
Get full text links from a DOI
}
\examples{
\donttest{
# pdf link
cr_full_GET(doi = "10.5555/515151", "pdf")

# xml and plain text links
out <- cr_works(filter=c(has_full_text = TRUE))
dois <- out$data$DOI
cr_full_GET(dois[1], "xml")
cr_full_GET(dois[1], "plain")
cr_full_GET(dois[1], "all")

# No links
cr_full_GET(doi = cr_r(1), "xml")

}
}

