% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_fundref.r
\name{cr_fundref}
\alias{cr_funders}
\alias{cr_fundref}
\title{Search the CrossRef Fundref API}
\usage{
cr_fundref(dois = NULL, query = NULL, filter = NULL, offset = NULL,
  limit = NULL, sample = NULL, sort = NULL, order = NULL,
  works = FALSE, cursor = NULL, cursor_max = 5000, .progress = "none",
  ...)

cr_funders(dois = NULL, query = NULL, filter = NULL, offset = NULL,
  limit = NULL, sample = NULL, sort = NULL, order = NULL,
  works = FALSE, cursor = NULL, cursor_max = 5000, .progress = "none",
  ...)
}
\arguments{
\item{dois}{Search by a single DOI or many DOIs.}

\item{query}{Query terms}

\item{filter}{Filter options. See details.}

\item{offset}{Number of record to start at, from 1 to infinity.}

\item{limit}{Number of results to return in the query. Not relavant when searching with specific
dois. Default: 20. Max: 1000}

\item{sample}{(integer) Number of random results to return. when you use the sample parameter, 
the rows and offset parameters are ignored. Ignored unless \code{works=TRUE}}

\item{sort}{(character) Field to sort on, one of score, relevance, updated, deposited, indexed, 
or published.}

\item{order}{(character) Sort order, one of 'asc' or 'desc'}

\item{works}{(logical) If TRUE, works returned as well, if not then not.}

\item{cursor}{(character) Cursor character string to do deep paging. Default is None.
Pass in '*' to start deep paging. Any combination of query, filters and facets may be
used with deep paging cursors. While rows may be specified along with cursor, offset
and sample cannot be used.
See https://github.com/CrossRef/rest-api-doc/blob/master/rest_api.md#deep-paging-with-cursors}

\item{cursor_max}{(integer) Max records to retrieve. Only used when cursor param used. 
Because deep paging can result in continuous requests until all are retrieved, use this
parameter to set a maximum number of records. Of course, if there are less records
found than this value, you will get only those found.}

\item{.progress}{Show a \code{plyr}-style progress bar? Options are "none", "text", 
"tk", "win", and "time".  See \code{\link[plyr]{create_progress_bar}} for details 
of each. Only used when passing in multiple ids (e.g., multiple DOIs, DOI prefixes,
etc.)}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Search the CrossRef Fundref API
}
\details{
BEWARE: The API will only work for CrossRef DOIs.

This function name changing to \code{cr_funders} in the next version - both work for now
}
\section{NOTE}{

Funders without IDs don't show up on the /funders route, and in this 
function. Some funders don't have assigned IDs in Crossref's system, 
so won't show up in searches.
}
\examples{
\dontrun{
cr_fundref(query="NSF", limit=1)
cr_fundref(query="NSF")
cr_fundref(dois='10.13039/100000001')
out <- cr_fundref(dois=c('10.13039/100000001','10.13039/100000015'))
out['10.13039/100000001']
out[['10.13039/100000001']]

cr_fundref(dois='10.13039/100000001')
cr_fundref(dois='10.13039/100000001', works=TRUE, limit=5)

cr_fundref(dois=c('10.13039/100000001','10.13039/100000015'))
cr_fundref(dois=c('10.13039/100000001','10.13039/100000015'), works=TRUE)

# Curl options
library('httr')
cr_fundref(dois='10.13039/100000001', config=verbose())

# If not found, and only 1 DOI given, list of NA elements returned
cr_fundref("adfadfaf")
# If not found, and > 1 DOI given, those not found dropped
cr_fundref(dois=c("adfadfaf","asfasf"))
cr_fundref(dois=c("adfadfaf","asfasf"), works=TRUE)
cr_fundref(dois=c("10.13039/100000001","asfasf"))
cr_fundref(dois=c("10.13039/100000001","asfasf"), works=TRUE)

# Use the cursor for deep paging
cr_funders('100000001', works = TRUE, cursor = "*", cursor_max = 500, limit = 100)
cr_funders(c('100000001', '100000002'), works = TRUE, cursor = "*",
   cursor_max = 300, limit = 100) 
}
}
\references{
\url{https://github.com/CrossRef/rest-api-doc/blob/master/rest_api.md}
}

