% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cr_fundref_funders}
\alias{cr_fundref_funders}
\title{Search the CrossRef Fundref API - the funders endpoint}
\usage{
cr_fundref_funders(funder_dois = NULL, query = NULL, filter = NULL,
  offset = NULL, limit = NULL, works = FALSE, ...)
}
\arguments{
\item{funder_dois}{Search by a single DOI or many DOIs.}

\item{format}{name of the format.}

\item{style}{a CSL style (for text format only)}

\item{locale}{language locale}

\item{...}{Named parameters passed on to httr::GET}
}
\description{
BEWARE: The API will only work for CrossRef DOIs.
}
\details{
See \url{bit.ly/1nIjfN5} for more info on the Fundref API service.
}
\examples{
\dontrun{
cr_fundref_funders(query="NSF")
cr_fundref_funders(query="NSF", limit=1)
cr_fundref_funders(funder_dois='10.13039/100000001')
out <- cr_fundref_funders(funder_dois=c('10.13039/100000001','10.13039/100000015'))
out['10.13039/100000001']
out[['10.13039/100000001']]

# Check the DOI minting agency
cr_agency(funder_dois='10.13039/100000001')

cr_fundref_funders(funder_dois='10.13039/100000001')
cr_fundref_funders(funder_dois='10.13039/100000001')

# Curl options
library('httr')
cr_fundref_funders(funder_dois='10.13039/100000001', config=verbose())
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

