% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{cr_fundref_members}
\alias{cr_fundref_members}
\title{Search the CrossRef Fundref API - the members endpoint}
\usage{
cr_fundref_members(member_ids = NULL, query = NULL, filter = NULL,
  offset = NULL, limit = NULL, sample = NULL, sort = NULL,
  order = NULL, facet = FALSE, works = FALSE, .progress = "none", ...)
}
\arguments{
\item{member_ids}{One or more member ids. See examples. ALternatively, you can query for them
using the query parameter.}

\item{query}{Query terms}

\item{filter}{Filter options. See details.}

\item{offset}{Number of record to start at, from 1 to infinity.}

\item{limit}{Number of results to return in the query. Not relavant when searching with specific
dois. Default: 20. Max: 1000}

\item{sample}{(integer) Number of random results to return. when you use the sample parameter,
the rows and offset parameters are ignored.}

\item{sort}{(character) Field to sort on, one of score, relevance, updated, deposited, indexed,
or published.}

\item{order}{(character) Sort order, one of 'asc' or 'desc'}

\item{facet}{(logical) Include facet results.}

\item{works}{(logical) If TRUE, works returned as well, if not then not.}

\item{...}{Named parameters passed on to httr::GET}

\item{progress}{Show a \code{plyr}-style progress bar? Options are "none", "text", "tk", "win, and "time".  See \link[pkg:plyr]{create_progress_bar} for details of each.}
}
\description{
BEWARE: The API will only work for CrossRef DOIs.
}
\details{
See \url{bit.ly/1nIjfN5} for more info on the Fundref API service.

See the filters vignette for options for filters.
}
\examples{
\dontrun{
cr_fundref_members(member_ids=98)
cr_fundref_members(query='hindawi')
out <- cr_fundref_members(member_ids=c(10,98))
out[['10']]$counts
out[['98']]$coverage

# curl options
library('httr')
cr_fundref_members(member_ids=98, config=verbose())
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

