% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cr_fundref_works}
\alias{cr_fundref_works}
\title{Search the CrossRef Fundref API - the works endpoint}
\usage{
cr_fundref_works(dois = NULL, query = NULL, filter = NULL,
  offset = NULL, limit = NULL, ...)
}
\arguments{
\item{dois}{Search by a single DOI or many DOIs.}

\item{format}{name of the format.}

\item{style}{a CSL style (for text format only)}

\item{locale}{language locale}

\item{...}{Named parameters passed on to httr::GET}
}
\description{
BEWARE: The API will only work for CrossRef DOIs.
}
\details{
See \url{bit.ly/1nIjfN5} for more info on the Fundref API service.
}
\examples{
\dontrun{
cr_fundref_works(query="NSF")
cr_fundref_works(dois='10.13039/100000001')
cr_fundref_works(dois='10.13039/100000001')
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

