% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cr_GET}
\alias{cr_GET}
\title{Crossref api internal handler}
\usage{
cr_GET(endpoint, args, todf = TRUE, ...)
}
\arguments{
\item{url}{endpoint}

\item{args}{Query args}

\item{...}{curl options}
}
\description{
Crossref api internal handler
}
\keyword{internal}

