% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cr_prefixes}
\alias{cr_prefixes}
\title{Search CrossRef prefixes}
\usage{
cr_prefixes(prefixes, query = NULL, filter = NULL, offset = NULL,
  limit = NULL, sample = NULL, sort = NULL, order = NULL,
  facet = FALSE, works = FALSE, .progress = "none", ...)
}
\arguments{
\item{prefixes}{(character) Publisher prefixes, one or more in a vector or list. Required.}

\item{query}{Query terms}

\item{filter}{Filter options. See details.}

\item{offset}{Number of record to start at, from 1 to infinity.}

\item{limit}{Number of results to return in the query. Not relavant when searching with specific
dois. Default: 20. Max: 1000}

\item{sample}{(integer) Number of random results to return. when you use the sample parameter,
the rows and offset parameters are ignored.}

\item{sort}{(character) Field to sort on, one of score, relevance, updated, deposited, indexed,
or published.}

\item{order}{(character) Sort order, one of 'asc' or 'desc'}

\item{facet}{(logical) Include facet results. Facet data not yet included in output.}

\item{.progress}{Show a \code{plyr}-style progress bar? Options are "none", "text", "tk", "win,
and "time".  See \link[plyr]{create_progress_bar} for details of each.}

\item{...}{Named parameters passed on to httr::GET}
}
\description{
BEWARE: The API will only work for CrossRef DOIs.
}
\details{
See \url{bit.ly/1nIjfN5} for more info on the Fundref API service.

See the filters vignette for options for filters.
}
\examples{
\dontrun{
cr_prefixes(prefixes="10.1016")
cr_prefixes(prefixes="10.1016", works=TRUE)
cr_prefixes(prefixes=c('10.1016','10.1371','10.1023','10.4176','10.1093'))
cr_prefixes(prefixes=c('10.1016','10.1371'), works=TRUE)
cr_works(prefixes="10.1016", filter=c(has_full_text=TRUE), limit=5)
cr_works(prefixes="10.1016", query='ecology', limit=4)
cr_works(prefixes="10.1016", query='ecology', limit=4)

# facets
## if works is FALSE, then facets ignored
cr_prefixes(prefixes="10.1016", works=FALSE, facet=TRUE)
cr_prefixes(prefixes="10.1016", works=TRUE, facet=TRUE)
cr_prefixes(prefixes=c('10.1016','10.1371'), works=TRUE, facet=TRUE)
}
}

