% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cr_r}
\alias{cr_r}
\title{Get a random set of DOI's through CrossRef.}
\usage{
cr_r(sample = 10, ...)
}
\arguments{
\item{sample}{The number of returned random DOIs. Maximum is 1000, default 20.
Note that a request for 1000 random DOIs will take a few seconds to
complete, whereas a request for 20 will take ~1 second.}

\item{...}{Further args passed on to \code{\link{cr_works}}}
}
\value{
A character vector of DOIs
}
\description{
Get a random set of DOI's through CrossRef.
}
\examples{
\dontrun{
# Default search gets 10 random DOIs
cr_r()
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\references{
\url{http://random.labs.crossref.org/}
}
\seealso{
\code{\link{cr_search}}, \code{\link{cr_citation}}, \code{\link{cr_search_free}}
}

