% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cr_search}
\alias{cr_search}
\title{Search the CrossRef Metatdata API.}
\usage{
cr_search(query = NULL, doi = NULL, page = NULL, rows = NULL,
  sort = NULL, year = NULL, type = NULL, ...)
}
\arguments{
\item{query}{Query terms.}

\item{doi}{Search by a single DOI or many DOIs.}

\item{page}{Page to return from results.}

\item{rows}{Number of records to return.}

\item{sort}{Sort either by "score" or "year".}

\item{year}{Year to search.}

\item{type}{Record type, e.g., "Journal Article" or "Journal Issue"}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Search the CrossRef Metatdata API.
}
\details{
See \url{http://search.labs.crossref.org/help/api} for more info on this
		Crossref API service.
}
\examples{
\donttest{
cr_search(query = c("renear", "palmer"))
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\seealso{
\code{\link{cr_r}}, \code{\link{cr_citation}}, \code{\link{cr_search_free}}
}

