% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cr_search2}
\alias{cr_search2}
\title{Search the CrossRef Metatdata API.}
\usage{
cr_search2(query, doi = NULL, page = NULL, rows = NULL, sort = NULL,
  year = NULL, type = NULL)
}
\arguments{
\item{query}{Query terms.}

\item{doi}{Search by a single DOI or many DOIs.}

\item{page}{Page to return from results.}

\item{rows}{Number of records to return.}

\item{sort}{Sort either by "score" or "year".}

\item{year}{Year to search.}

\item{type}{Record type, e.g., "Journal Article" or "Journal Issue"}
}
\description{
Search the CrossRef Metatdata API.
}
\details{
See \url{https://github.com/CrossRef/rest-api-doc} for more info on this
  	Crossref API service.
}
\examples{
\dontrun{
cr_search2(query = c("renear", "palmer"))
}
}

