% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_search_free.r
\name{cr_search_free}
\alias{cr_search_free}
\title{Search the CrossRef Metadata for DOIs using free form references.}
\usage{
cr_search_free(...)
}
\description{
Search the CrossRef Metadata for DOIs using free form references.
}
\keyword{internal}
