% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cr_search_free}
\alias{cr_search_free}
\title{Search the CrossRef Metatdata for DOIs using free form references.}
\usage{
cr_search_free(query, url = "http://search.crossref.org/links")
}
\arguments{
\item{query}{Reference query; a character vector of length 1 or greater,
comma-separated of course.}

\item{url}{Base url for the Crossref metadata API.}
}
\description{
Search the CrossRef Metatdata for DOIs using free form references.
}
\details{
Have to have at least three terms in each search query.
}
\examples{
\donttest{
# search with title, author, year, and journal
cr_search_free(query = "Piwowar Sharing Detailed Research Data Is Associated with
Increased Citation Rate PLOS one 2007")
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\seealso{
\code{\link{cr_search}}, \code{\link{cr_r}}, \code{\link{cr_citation}}
}

