% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cr_works_links.R
\name{cr_works_links}
\alias{cr_works_links}
\title{Get text mining links via the CrossRef works endpoint}
\usage{
cr_works_links(dois = NULL, ...)
}
\arguments{
\item{dois}{One or more DOIs.}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Get text mining links via the CrossRef works endpoint
}
\examples{
\dontrun{
# elife
out <- cr_members(4374, filter=c(has_full_text = TRUE), works = TRUE)
dois <- out$data$DOI[1:5]
cr_works_links(dois)

# hindawi
out <- cr_members(98, filter=c(has_full_text = TRUE), works = TRUE)
dois <- out$data$DOI[1:5]
cr_works_links(dois)

cr_works_links(dois = "10.1016/j.fm.2004.01.001")
dois <- c("10.1016/s0747-7171(03)00168-8","10.1016/j.fm.2004.01.002","10.1016/j.fm.2004.01.007")
}
}

