% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crosscite.R
\name{crosscite}
\alias{crosscite}
\title{Crosscite - citation formatter}
\usage{
crosscite(dois, style = "apa", locale = "en-US", .progress = "none", ...)
}
\arguments{
\item{dois}{Search by a single DOI or many DOIs.}

\item{style}{a CSL style (for text format only). See \code{\link{get_styles}}
for options. Default: apa. If there's a style that CrossRef doesn't support you'll get a
\code{(500) Internal Server Error}}

\item{locale}{Language locale. See \code{?Sys.getlocale}}

\item{.progress}{Show a \code{plyr}-style progress bar? Options are "none", "text", 
"tk", "win", and "time".  See \code{\link[plyr]{create_progress_bar}} for details 
of each. Only used when passing in multiple ids (e.g., multiple DOIs, DOI prefixes,
etc.)}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Crosscite - citation formatter
}
\details{
See \url{http://www.crosscite.org/cn/} for more info on the
  	Crossref Content Negotiation API service.

This function is now deprecated. It will be removed in the next version
of this package. Use \code{\link{cr_cn}} instead.
}
\examples{
\dontrun{
crosscite("10.5284/1011335")
crosscite(c('10.5169/SEALS-52668','10.2314/GBV:493109919','10.2314/GBV:493105263',
   '10.2314/GBV:487077911','10.2314/GBV:607866403'))
}
}

