% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdf.R
\name{extract_xpdf}
\alias{extract_xpdf}
\title{Extract text from a single pdf document.}
\usage{
extract_xpdf(...)
}
\description{
Extract text from a single pdf document.
}
\note{
see \code{crminer::crm_extract}
}
\keyword{internal}
