% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\alias{filter_names}
\alias{filter_details}
\title{Get filter details and names.}
\usage{
filter_names()

filter_details(x = NULL)
}
\arguments{
\item{x}{(character) Optional filter name. If not given, all filters
returned.}
}
\description{
Get filter details and names.
}
\details{
Note that all filter names in this package have dashes replaced
with underscores as dashes cause problems in an R console.
}
\examples{
filter_names()
filter_details()
filter_details()$assertion
filter_details()$assertion$possible_values
filter_details()$assertion$description
filter_details("issn")
filter_details("iss")
filter_details(c("issn", "alternative_id"))
}
