% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{fundref_GET}
\alias{fundref_GET}
\title{Fundref api internal handler}
\usage{
fundref_GET(endpoint, args, ...)
}
\arguments{
\item{args}{Query args}

\item{...}{curl options}

\item{url}{endpoint}
}
\description{
Fundref api internal handler
}
\keyword{internal}

