% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_cn.r
\name{GET_agency_id}
\alias{GET_agency_id}
\title{Get doi agency id to identify resource location}
\usage{
GET_agency_id(x, ...)
}
\arguments{
\item{x}{doi}
}
\description{
Get doi agency id to identify resource location
}
\keyword{internal}

