% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_md_plus_token}
\alias{get_md_plus_token}
\title{Retrieve Optional \href{https://www.crossref.org/services/metadata-retrieval/metadata-plus/}{Metadata Plus} Token}
\usage{
get_md_plus_token()
}
\description{
Retrieves the secret token to authenticate against the paid-for Metadata Plus service.
}
\details{
To use your paid-for Metadata Plus API access,
save your secret token as an environment variable named \code{crossref_plus}.

You can declare such R-specific environment variables in your (user) \code{.Renviron}.
Open it with file.edit("~/.Renviron")`, add\preformatted{crossref_plus = YOUR_SECRET_TOKEN
}

then save the file and restart your R session.

Remember never to commit or share your \code{.Renviron} file, customarily used
for storing secrets such as this.

Please be aware that without valid authentication token,
requests will cause 401 HTTP errors.
}
