% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_styles}
\alias{get_styles}
\title{Get list of styles from github.com/citation-style-language/styles}
\usage{
get_styles(...)
}
\arguments{
\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Get list of styles from github.com/citation-style-language/styles
}
\examples{
\donttest{
get_styles()[1:5]
}
}

