% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pmid2doi}
\alias{doi2pmid}
\alias{pmid2doi}
\title{Get a PMID from a DOI, and vice versa.}
\usage{
pmid2doi(x, simplify = FALSE, ...)

doi2pmid(x, simplify = FALSE, ...)
}
\arguments{
\item{x}{(character) One or more DOIs or PMIDs}

\item{simplify}{(logical) Whether to simplify result to vector.}

\item{...}{Curl args passed on to \code{\link[httr]{GET}}.}
}
\description{
Get a PMID from a DOI, and vice versa.
}
\examples{
\donttest{
# get a pmid from a doi
doi2pmid("10.1016/0006-2944(75)90147-7")
}
}
\references{
Uses the \url{http://www.pmid2doi.org/} REST API.
}

