% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{rcrossref-deprecated}
\alias{rcrossref-deprecated}
\title{Deprecated functions in rcrossref}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{cr_search}}: The functionality of this function can 
 be achieved with the new Crossref API. See functions \code{\link{cr_works}} 
 et al.
 \item \code{\link{cr_search_free}}: The functionality of this function can 
 be achieved with the new Crossref API. See functions \code{\link{cr_works}} 
 et al.
 \item \code{\link{crosscite}}: The functionality of this function can be 
 achieved with \code{\link{cr_cn}}
}
}

