% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cr_full.R
\name{cr_ft_links}
\alias{cr_ft_links}
\alias{cr_ft_text}
\title{Get full text links from a DOI}
\usage{
cr_ft_links(doi, type = "xml", ...)

cr_ft_text(url, type = "xml", path = "~/.crossref", overwrite = TRUE,
  read = TRUE, verbose = TRUE, ...)
}
\arguments{
\item{doi}{A DOI}

\item{type}{One of xml, plain, pdf, or all}

\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Get full text links from a DOI
}
\examples{
\donttest{
# pdf link
cr_ft_links(doi = "10.5555/515151", "pdf")

# xml and plain text links
out <- cr_works(filter=c(has_full_text = TRUE))
dois <- out$data$DOI
cr_ft_links(dois[2], "xml")
cr_ft_links(dois[1], "plain")
cr_ft_links(dois[1], "all")

# No links
cr_ft_links(doi = cr_r(1), "xml")

# get full text
## elife
out <- cr_members(4374, filter=c(has_full_text = TRUE), works = TRUE)
### pdf
links <- cr_ft_links(out$data$DOI[1], "all")
cr_ft_text(links, read=FALSE)
cr_ft_text(links)
### xml
links <- cr_ft_links(out$data$DOI[1], "all")
cr_ft_text(links, 'xml')

## pensoft
out <- cr_members(2258, filter=c(has_full_text = TRUE), works = TRUE)
### pdf
links <- cr_ft_links(out$data$DOI[1], "all")
cr_ft_text(links, "pdf", read=FALSE)
cr_ft_text(links, "pdf")
### xml
cr_ft_text(links, 'xml')

## hindawi
out <- cr_members(98, filter=c(has_full_text = TRUE), works = TRUE)
### pdf
links <- cr_ft_links(out$data$DOI[1], "all")
cr_ft_text(links, "pdf", read=FALSE)
cr_ft_text(links, "pdf")
### xml
cr_ft_text(links, 'xml')

## search for works with full text, and with CC-BY 3.0 license
### you can see available licenses with cr_licenses() function
out <-
 cr_works(filter = list(has_full_text = TRUE,
   license_url="http://creativecommons.org/licenses/by/3.0/"))
links <- cr_ft_links(out$data$DOI[10], "all")
cr_ft_text(links, 'xml')

## elsevier - they don't actually give full text, ha ha, jokes on us!
out <- cr_members(78, filter=c(has_full_text = TRUE), works = TRUE)
links <- cr_ft_links(out$data$DOI[1], "all")
cr_ft_text(links, 'xml') # notice how this is just metadata
}
}
\keyword{internal}

