% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cr_fundref_members}
\alias{cr_fundref_members}
\title{Search the CrossRef Fundref API - the members endpoint}
\usage{
cr_fundref_members(member_ids = NULL, query = NULL, filter = NULL,
  offset = NULL, limit = NULL, ...)
}
\arguments{
\item{dois}{Search by a single DOI or many DOIs.}

\item{format}{name of the format.}

\item{style}{a CSL style (for text format only)}

\item{locale}{language locale}

\item{...}{Named parameters passed on to httr::GET}
}
\description{
BEWARE: The API will only work for CrossRef DOIs.
}
\details{
See \url{bit.ly/1nIjfN5} for more info on the Fundref API service.
}
\examples{
\dontrun{
cr_fundref_members(member_ids='10.13039/100000001')
out <- cr_fundref_members(member_ids=c('10.13039/100000001','10.13039/100000015'))
out['10.13039/100000001']
out[['10.13039/100000001']]
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}

