% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cr_r.r
\name{cr_r}
\alias{cr_r}
\title{Get a random set of DOI's through CrossRef.}
\usage{
cr_r(sample = 10, ...)
}
\arguments{
\item{sample}{The number of returned random DOIs. Maximum: 100. Default: 20.}

\item{...}{Further args passed on to \code{\link[=cr_works]{cr_works()}}}
}
\value{
A character vector of DOIs
}
\description{
Get a random set of DOI's through CrossRef.
}
\examples{
\dontrun{
# Default search gets 10 random DOIs
cr_r()

# Get 30 DOIs
cr_r(30)
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
