% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pdf.R
\name{extract_xpdf}
\alias{extract_xpdf}
\title{Extract text from a single pdf document.}
\usage{
extract_xpdf(path)
}
\arguments{
\item{path}{Path to a file}
}
\value{
An object of class xpdf_char
}
\description{
Extract text from a single pdf document.
}
\details{
Download xpdf from \url{http://www.foolabs.com/xpdf/download.html}
}
\examples{
\dontrun{
path <- system.file("examples", "MairChamberlain2014RJournal.pdf", package = "rcrossref")
(res_xpdf <- extract_xpdf(path))
res_xpdf$meta
res_xpdf$data
}
}

