% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_styles.R
\name{get_styles}
\alias{get_styles}
\title{Get list of styles from github.com/citation-style-language/styles}
\usage{
get_styles(...)
}
\arguments{
\item{...}{Named parameters passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Get list of styles from github.com/citation-style-language/styles
}
\examples{
\dontrun{
x <- get_styles()
x[1:10]
}
}
