% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get_styles.R
\name{get_styles}
\alias{get_styles}
\title{Get list of styles from github.com/citation-style-language/styles}
\usage{
get_styles(...)
}
\arguments{
\item{...}{Named parameters passed on to \code{\link[httr]{GET}}}
}
\description{
Get list of styles from github.com/citation-style-language/styles
}
\examples{
\dontrun{
get_styles()[1:5]
}
}

