% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/defunct.R
\name{rcrossref-defunct}
\alias{rcrossref-defunct}
\title{Defunct functions in rcrossref}
\description{
\itemize{
 \item \code{\link{cr_citation}}: Crossref is trying to sunset their OpenURL API, which
 this function uses. So this function is now removed. See the function \code{\link{cr_cn}},
 which does the same things, but with more functionality, using the new Crossref API.
}
}
\details{
The above function will be removed in a future version of this package.
}

