% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{rcrossref-defunct}
\alias{rcrossref-defunct}
\title{Defunct functions in rcrossref}
\description{
These functions are gone, no longer available.
}
\details{
\itemize{
\item \code{\link[=cr_citation]{cr_citation()}}: Crossref is trying to sunset their
OpenURL API, which  this function uses. So this function is now removed.
See the function \code{\link[=cr_cn]{cr_cn()}}, which does the same things, but with
more functionality, using the new Crossref API.
\item \code{\link[=pmid2doi]{pmid2doi()}} and \code{\link[=doi2pmid]{doi2pmid()}}: The API behind
these functions is down for good, see \code{\link[=id_converter]{id_converter()}} for
similar functionality.
\item \code{\link[=cr_search]{cr_search()}}: The functionality of this function can
be achieved with the new Crossref API. See functions \code{\link[=cr_works]{cr_works()}}
et al.
\item \code{\link[=cr_search_free]{cr_search_free()}}: The functionality of this function can
be achieved with the new Crossref API. See functions \code{\link[=cr_works]{cr_works()}}
et al.
\item \code{\link[=crosscite]{crosscite()}}: The functionality of this function can be
achieved with \code{\link[=cr_cn]{cr_cn()}}
\item \code{\link[=cr_fundref]{cr_fundref()}}: Crossref changed their name "fundref"
to "funders", so we've changed our function, see \code{\link[=cr_funders]{cr_funders()}}
\item \code{\link[=cr_ft_text]{cr_ft_text()}}: This function and other text mining
functions are incorporated in a new package \code{crminer}.
\item \code{\link[=cr_ft_links]{cr_ft_links()}}: This function and other text mining
functions are incorporated in a new package \code{crminer}.
\item \code{\link[=cr_ft_pdf]{cr_ft_pdf()}}: This function and other text mining
functions are  incorporated in a new package \code{crminer}.
\item \code{\link[=cr_ft_plain]{cr_ft_plain()}}: This function and other text mining
functions are incorporated in a new package \code{crminer}.
\item \code{\link[=cr_ft_text]{cr_ft_text()}}: This function and other text mining
functions are incorporated in a new package \code{crminer}.
\item \code{\link[=cr_ft_xml]{cr_ft_xml()}}: This function and other text mining
functions are incorporated in a new package \code{crminer}.
\item \code{\link[=as.tdmurl]{as.tdmurl()}}: This function and other text mining
functions are incorporated in a new package \code{crminer}.
\item \code{\link[=extract_xpdf]{extract_xpdf()}}: This function and other text mining
functions are incorporated in a new package \code{crminer}.
}
}
