% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rcrossref-deprecated}
\alias{rcrossref-deprecated}
\title{Deprecated functions in rcrossref}
\description{
\itemize{
 \item \code{\link{cr_citation}}: Crossref is trying to sunset their OpenURL API, which
 this function uses. So this function will be removed in a future version of this
 package. See the function \code{\link{cr_cn}}, which does the same things, but with
 more functionality, using the new Crossref API.
}
}
\details{
The above function will be removed in a future version of this package.
}

