% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_defunct.R
\name{rcrossref-deprecated}
\alias{rcrossref-deprecated}
\title{Deprecated functions in rcrossref}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
 \item \code{\link{cr_ft_text}}: This function and other text mining
 functions are being incorporated in a new package \code{crminer}.
 \item \code{\link{cr_ft_links}}: This function and other text mining
 functions are being incorporated in a new package \code{crminer}.
 \item \code{\link{cr_ft_pdf}}: This function and other text mining
 functions are being incorporated in a new package \code{crminer}.
 \item \code{\link{cr_ft_plain}}: This function and other text mining
 functions are being incorporated in a new package \code{crminer}.
 \item \code{\link{cr_ft_text}}: This function and other text mining
 functions are being incorporated in a new package \code{crminer}.
 \item \code{\link{cr_ft_xml}}: This function and other text mining
 functions are being incorporated in a new package \code{crminer}.
 \item \code{\link{as.tdmurl}}: This function and other text mining
 functions are being incorporated in a new package \code{crminer}.
 \item \code{\link{extract_xpdf}}: This function and other text mining
 functions are being incorporated in a new package \code{crminer}.
}
}

