% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rcrossref-package}
\alias{rcrossref}
\alias{rcrossref-package}
\title{R Client for Various CrossRef APIs.}
\description{
R Client for Various CrossRef APIs.
}
\details{
rcrossref interacts with the main Crossref metadata search API at
\url{https://github.com/CrossRef/rest-api-doc/blob/master/rest_api.md}, the old metadata
search API at \url{http://search.labs.crossref.org/}, the CrossRef OpenURL service
\url{http://www.crossref.org/openurl/}, and their DOI Content Negotiation  service at
\url{http://www.crosscite.org/cn/}.
}
\keyword{package}

