% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rcrossref-package.R
\docType{package}
\name{rcrossref-package}
\alias{rcrossref}
\alias{rcrossref-package}
\title{R Client for Various CrossRef APIs.}
\description{
R Client for Various CrossRef APIs.
}
\details{
rcrossref interacts with the main Crossref metadata search API at
\url{https://github.com/CrossRef/rest-api-doc/blob/master/rest_api.md}, the old metadata
search API at \url{http://search.labs.crossref.org/}, their DOI Content Negotiation
service at \url{http://www.crosscite.org/cn/}, and the \emph{Text and Data Mining}
project \url{http://tdmsupport.crossref.org/}.

Note that functions that use the OpenURL service are now deprecated, and will be defunct
soon.
}
\keyword{package}

